backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentflow-backend"
version = "0.1.0"
description = "FastAPI backend baseline for AgentFlow"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "alembic>=1.14.0",
  "fastapi>=0.115.6",
  "pydantic-settings>=2.7.1",
  "psycopg[binary]>=3.2.3",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.36",
  "uvicorn[standard]>=0.34.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "mypy>=1.14.1",
  "pytest>=8.3.4",
  "ruff>=0.9.2"
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = ["B", "E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["app"]
exclude = ["alembic"]
